Struct eth_bridge::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn http_request( url: &str, body: Vec<u8>, headers: &[(&'static str, String)] ) -> Result<Vec<u8>, Error<T>>

Makes off-chain HTTP request.

pub fn json_rpc_request<I: Serialize, O: for<'de> Deserialize<'de>>( url: &str, id: u64, method: &str, params: &I, headers: &[(&'static str, String)] ) -> Result<O, Error<T>>

Makes JSON-RPC request.

pub fn eth_json_rpc_request<I: Serialize, O: for<'de> Deserialize<'de>>( method: &str, params: &I, network_id: T::NetworkId ) -> Result<O, Error<T>>

Makes request to a Sidechain node. The node URL and credentials are stored in the local storage.

pub fn substrate_json_rpc_request<I: Serialize, O: for<'de> Deserialize<'de>>( method: &str, params: &I ) -> Result<O, Error<T>>

Makes request to the local node. The node URL is stored in the local storage.

pub fn load_is_used( hash: H256, network_id: T::NetworkId ) -> Result<bool, Error<T>>

Queries Sidechain’s contract variable used.

pub fn load_current_height(network_id: T::NetworkId) -> Result<u64, Error<T>>

Queries current height of Sidechain.

pub fn load_tx( hash: H256, network_id: T::NetworkId ) -> Result<Transaction, Error<T>>

Loads a Sidechain transaction by the hash and ensures that it came from a known contract.

pub fn load_tx_receipt( hash: H256, network_id: T::NetworkId ) -> Result<TransactionReceipt, Error<T>>

Loads a Sidechain transaction receipt by the hash and ensures that it came from a known contract.

pub fn load_substrate_finalized_header( ) -> Result<SubstrateHeaderLimited, Error<T>>where T: CreateSignedTransaction<<T as Config>::RuntimeCall>,

Queries the current finalized block of the local node with chain_getFinalizedHead and chain_getHeader RPC calls.

pub fn load_substrate_block( number: T::BlockNumber ) -> Result<SubstrateBlockLimited, Error<T>>where T: CreateSignedTransaction<<T as Config>::RuntimeCall>,

Queries a block at the given height of the local node with chain_getBlockHash and chain_getBlock RPC calls.

pub fn load_transfers_logs( network_id: T::NetworkId, from_block: u64, to_block: u64 ) -> Result<Vec<Log>, Error<T>>

Queries the sidechain node for the transfer logs emitted within from_block and to_block.

Uses the eth_getLogs method with a filter on log topic.

impl<T: Config> Pallet<T>

pub fn parse_main_event( network_id: T::NetworkId, logs: &[Log], kind: IncomingTransactionRequestKind ) -> Result<ContractEvent<H160, T::AccountId, Balance>, Error<T>>

Loops through the given array of logs and finds the first one that matches the type and topic.

pub fn parse_old_incoming_request_method_call( incoming_request: LoadIncomingTransactionRequest<T>, tx: Transaction ) -> Result<IncomingRequest<T>, Error<T>>

Tries to parse a method call on one of old Sora contracts (XOR and VAL).

Since the XOR and VAL contracts don’t have the same interface and events that the modern bridge contract have, and since they can’t be changed we have to provide a special parsing flow for some of the methods that we might use.

pub fn ensure_known_contract( to: H160, network_id: T::NetworkId ) -> Result<(), Error<T>>

Checks that the given contract address is known to the bridge network.

There are special cases for XOR and VAL contracts.

impl<T: Config> Pallet<T>

pub fn get_requests( hashes: &[H256], network_id: Option<T::NetworkId>, redirect_finished_load_requests: bool ) -> Result<Vec<(OffchainRequest<T>, RequestStatus)>, DispatchError>

Get requests data and their statuses by hash.

pub fn get_approved_requests( hashes: &[H256], network_id: Option<T::NetworkId> ) -> Result<Vec<(OutgoingRequestEncoded, Vec<SignatureParams>)>, DispatchError>

Get approved outgoing requests data and proofs.

pub fn get_approvals( hashes: &[H256], network_id: Option<T::NetworkId> ) -> Result<Vec<Vec<SignatureParams>>, DispatchError>

Get requests approvals.

pub fn get_account_requests( account: &T::AccountId, status_filter: Option<RequestStatus> ) -> Result<Vec<(T::NetworkId, H256)>, DispatchError>

Get account requests list.

pub fn get_registered_assets( network_id: Option<T::NetworkId> ) -> Result<Vec<(AssetKind, (<T as Config>::AssetId, BalancePrecision), Option<(H160, BalancePrecision)>)>, DispatchError>

Get registered assets and tokens.

impl<T: Config> Pallet<T>

pub fn is_peer(who: &T::AccountId, network_id: T::NetworkId) -> bool

Checks if the account is a bridge peer.

pub fn convert_precision( precision_from: BalancePrecision, precision_to: BalancePrecision, amount: Balance ) -> Result<(Balance, Balance), Error<T>>

Converts amount from one precision to another and and returns it with a difference of the amounts. It also checks that no information was lost during multiplication, otherwise returns an error.

impl<T: Config> Pallet<T>

pub fn register_bridge( origin: OriginFor<T>, bridge_contract_address: H160, initial_peers: Vec<T::AccountId>, signature_version: BridgeSignatureVersion ) -> DispatchResultWithPostInfo

Register a new bridge.

Parameters:

bridge_contract_address - address of smart-contract deployed on a corresponding network.

initial_peers - a set of initial network peers.

pub fn add_asset( origin: OriginFor<T>, asset_id: <T as Config>::AssetId, network_id: <T as Config>::NetworkId ) -> DispatchResultWithPostInfo

Add a Thischain asset to the bridge whitelist.

Can only be called by root.

Parameters:

asset_id - Thischain asset identifier.

network_id - network identifier to which the asset should be added.

pub fn add_sidechain_token( origin: OriginFor<T>, token_address: H160, symbol: String, name: String, decimals: u8, network_id: <T as Config>::NetworkId ) -> DispatchResultWithPostInfo

Add a Sidechain token to the bridge whitelist.

Parameters:

token_address - token contract address.

symbol - token symbol (ticker).

name - token name.

decimals - token precision.

network_id - network identifier.

pub fn transfer_to_sidechain( origin: OriginFor<T>, asset_id: <T as Config>::AssetId, to: H160, amount: Balance, network_id: <T as Config>::NetworkId ) -> DispatchResultWithPostInfo

Transfer some amount of the given asset to Sidechain address.

Note: if the asset kind is Sidechain, the amount should fit in the asset’s precision on sidechain (SidechainAssetPrecision) without extra digits. For example, assume some ERC-20 (T) token has decimals=6, and the corresponding asset on substrate has 7. Alice’s balance on thischain is 0.1000009. If Alice would want to transfer all the amount, she will get an error NonZeroDust, because of the 9 at the end, so, the correct amount would be 0.100000 (only 6 digits after the decimal point).

Parameters:

asset_id - thischain asset id.

to - sidechain account id.

amount - amount of the asset.

network_id - network identifier.

pub fn request_from_sidechain( origin: OriginFor<T>, eth_tx_hash: H256, kind: IncomingRequestKind, network_id: <T as Config>::NetworkId ) -> DispatchResultWithPostInfo

Load incoming request from Sidechain by the given transaction hash.

Parameters:

eth_tx_hash - transaction hash on Sidechain.

kind - incoming request type.

network_id - network identifier.

pub fn finalize_incoming_request( origin: OriginFor<T>, hash: H256, network_id: <T as Config>::NetworkId ) -> DispatchResultWithPostInfo

Finalize incoming request (see Pallet::finalize_incoming_request_inner).

Can be only called from a bridge account.

Parameters:

request - an incoming request.

network_id - network identifier.

pub fn add_peer( origin: OriginFor<T>, account_id: T::AccountId, address: H160, network_id: <T as Config>::NetworkId ) -> DispatchResultWithPostInfo

Add a new peer to the bridge peers set.

Parameters:

account_id - account id on thischain.

address - account id on sidechain.

network_id - network identifier.

pub fn remove_peer( origin: OriginFor<T>, account_id: T::AccountId, peer_address: Option<H160>, network_id: <T as Config>::NetworkId ) -> DispatchResultWithPostInfo

Remove peer from the the bridge peers set.

Parameters:

account_id - account id on thischain.

network_id - network identifier.

pub fn prepare_for_migration( origin: OriginFor<T>, network_id: <T as Config>::NetworkId ) -> DispatchResultWithPostInfo

Prepare the given bridge for migration.

Can only be called by an authority.

Parameters:

network_id - bridge network identifier.

pub fn migrate( origin: OriginFor<T>, new_contract_address: H160, erc20_native_tokens: Vec<H160>, network_id: <T as Config>::NetworkId, new_signature_version: BridgeSignatureVersion ) -> DispatchResultWithPostInfo

Migrate the given bridge to a new smart-contract.

Can only be called by an authority.

Parameters:

new_contract_address - new sidechain ocntract address.

erc20_native_tokens - migrated assets ids.

network_id - bridge network identifier.

pub fn register_incoming_request( origin: OriginFor<T>, incoming_request: IncomingRequest<T> ) -> DispatchResultWithPostInfo

Register the given incoming request and add it to the queue.

Calls validate and prepare on the request, adds it to the queue and maps it with the corresponding load-incoming-request and removes the load-request from the queue.

Can only be called by a bridge account.

pub fn import_incoming_request( origin: OriginFor<T>, load_incoming_request: LoadIncomingRequest<T>, incoming_request_result: Result<IncomingRequest<T>, DispatchError> ) -> DispatchResultWithPostInfo

Import the given incoming request.

Register’s the load request, then registers and finalizes the incoming request if it succeeded, otherwise aborts the load request.

Can only be called by a bridge account.

pub fn approve_request( origin: OriginFor<T>, ocw_public: Public, hash: H256, signature_params: SignatureParams, network_id: <T as Config>::NetworkId ) -> DispatchResultWithPostInfo

Approve the given outgoing request. The function is used by bridge peers.

Verifies the peer signature of the given request and adds it to RequestApprovals. Once quorum is collected, the request gets finalized and removed from request queue.

pub fn abort_request( origin: OriginFor<T>, hash: H256, error: DispatchError, network_id: <T as Config>::NetworkId ) -> DispatchResultWithPostInfo

Cancels a registered request.

Loads request by the given hash, cancels it, changes its status to Failed and removes it from the request queues.

Can only be called from a bridge account.

pub fn force_add_peer( origin: OriginFor<T>, who: T::AccountId, address: H160, network_id: <T as Config>::NetworkId ) -> DispatchResultWithPostInfo

Add the given peer to the peers set without additional checks.

Can only be called by a root account.

pub fn remove_sidechain_asset( origin: OriginFor<T>, asset_id: <T as Config>::AssetId, network_id: <T as Config>::NetworkId ) -> DispatchResultWithPostInfo

Remove asset

Can only be called by root.

pub fn register_existing_sidechain_asset( origin: OriginFor<T>, asset_id: <T as Config>::AssetId, token_address: H160, network_id: <T as Config>::NetworkId ) -> DispatchResultWithPostInfo

Register existing asset

Can only be called by root.

impl<T: Config> Pallet<T>

pub fn requests_queue<KArg>(k: KArg) -> Vec<H256> where KArg: EncodeLike<<T as Config>::NetworkId>,

Registered requests queue handled by off-chain workers.

impl<T: Config> Pallet<T>

pub fn request<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> Option<OffchainRequest<T>>where KArg1: EncodeLike<<T as Config>::NetworkId>, KArg2: EncodeLike<H256>,

Registered requests.

impl<T: Config> Pallet<T>

pub fn load_to_incoming_request_hash<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> H256where KArg1: EncodeLike<<T as Config>::NetworkId>, KArg2: EncodeLike<H256>,

Used to identify an incoming request by the corresponding load request.

impl<T: Config> Pallet<T>

pub fn request_status<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> Option<RequestStatus>where KArg1: EncodeLike<<T as Config>::NetworkId>, KArg2: EncodeLike<H256>,

Requests statuses.

impl<T: Config> Pallet<T>

pub fn request_submission_height<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> T::BlockNumberwhere KArg1: EncodeLike<<T as Config>::NetworkId>, KArg2: EncodeLike<H256>,

Requests submission height map (on substrate).

impl<T: Config> Pallet<T>

pub fn approvals<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> BTreeSet<SignatureParams>where KArg1: EncodeLike<<T as Config>::NetworkId>, KArg2: EncodeLike<H256>,

Outgoing requests approvals.

impl<T: Config> Pallet<T>

pub fn account_requests<KArg>(k: KArg) -> Vec<(<T as Config>::NetworkId, H256)> where KArg: EncodeLike<T::AccountId>,

Requests made by an account.

impl<T: Config> Pallet<T>

pub fn registered_asset<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> Option<AssetKind>where KArg1: EncodeLike<<T as Config>::NetworkId>, KArg2: EncodeLike<T::AssetId>,

Registered asset kind.

impl<T: Config> Pallet<T>

pub fn sidechain_asset_precision<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> BalancePrecisionwhere KArg1: EncodeLike<<T as Config>::NetworkId>, KArg2: EncodeLike<T::AssetId>,

Precision (decimals) of a registered sidechain asset. Should be the same as in the ERC-20 contract.

impl<T: Config> Pallet<T>

pub fn registered_sidechain_asset<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> Option<T::AssetId>where KArg1: EncodeLike<<T as Config>::NetworkId>, KArg2: EncodeLike<H160>,

Registered token AssetId on Thischain.

impl<T: Config> Pallet<T>

pub fn registered_sidechain_token<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> Option<H160>where KArg1: EncodeLike<<T as Config>::NetworkId>, KArg2: EncodeLike<T::AssetId>,

Registered asset address on Sidechain.

impl<T: Config> Pallet<T>

pub fn peers<KArg>(k: KArg) -> BTreeSet<T::AccountId>where KArg: EncodeLike<<T as Config>::NetworkId>,

Network peers set.

impl<T: Config> Pallet<T>

pub fn pending_peer<KArg>(k: KArg) -> Option<T::AccountId>where KArg: EncodeLike<<T as Config>::NetworkId>,

Network pending (being added/removed) peer.

impl<T: Config> Pallet<T>

pub fn peer_account_id<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> Option<T::AccountId>where KArg1: EncodeLike<<T as Config>::NetworkId>, KArg2: EncodeLike<H160>,

Peer account ID on Thischain.

impl<T: Config> Pallet<T>

pub fn peer_address<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> H160where KArg1: EncodeLike<<T as Config>::NetworkId>, KArg2: EncodeLike<T::AccountId>,

Peer address on Sidechain.

impl<T: Config> Pallet<T>

pub fn bridge_account<KArg>(k: KArg) -> Option<T::AccountId>where KArg: EncodeLike<<T as Config>::NetworkId>,

Multi-signature bridge peers’ account. None if there is no account and network with the given ID.

impl<T: Config> Pallet<T>

pub fn authority_account() -> Option<T::AccountId>

Thischain authority account.

impl<T: Config> Pallet<T>

pub fn bridge_contract_status<KArg>(k: KArg) -> Option<BridgeStatus>where KArg: EncodeLike<<T as Config>::NetworkId>,

Bridge status.

impl<T: Config> Pallet<T>

pub fn bridge_contract_address<KArg>(k: KArg) -> H160where KArg: EncodeLike<<T as Config>::NetworkId>,

Smart-contract address on Sidechain.

impl<T: Config> Pallet<T>

pub fn xor_master_contract_address() -> H160

Sora XOR master contract address.

impl<T: Config> Pallet<T>

pub fn val_master_contract_address() -> H160

Sora VAL master contract address.

impl<T: Config> Pallet<T>

pub fn bridge_signature_version<KArg>(k: KArg) -> BridgeSignatureVersionwhere KArg: EncodeLike<<T as Config>::NetworkId>,

impl<T: Config> Pallet<T>

pub fn pending_bridge_signature_version<KArg>( k: KArg ) -> Option<BridgeSignatureVersion>where KArg: EncodeLike<<T as Config>::NetworkId>,