Module eth_bridge::types

Web3 Types

Structs

Block

The block type returned from RPC calls. This is generic over a TX type.

BlockHeader

The block header type returned from RPC calls.

Bytes

Raw bytes wrapper

CallRequest

Call contract request (eth_call / eth_estimateGas)

Filter

Filter

FilterBuilder

Filter Builder

H64

H128

Fixed-size uninterpreted hash type with 16 bytes (128 bits) size.

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

H512

Fixed-size uninterpreted hash type with 64 bytes (512 bits) size.

H520

H2048

Bloom hash type with 256 bytes (2048 bits) size.

Log

A log produced by a transaction.

RawTransaction

Raw bytes of a signed, but not yet sent transaction

SubstrateBlockLimited

SubstrateHeaderLimited

SubstrateSignedBlockLimited

Transaction

Description of a Transaction, pending or in the chain.

TransactionReceipt

“Receipt” of an executed transaction: details of its execution.

TransactionRequest

Send Transaction Parameters

U64

Little-endian large integer type Unsigned 64-bit integer.

U128

Little-endian large integer type 128-bit unsigned integer.

U256

Little-endian large integer type 256-bit unsigned integer.

Enums

BlockId

Block Identifier

BlockNumber

Block Number

TransactionCondition

Represents condition on minimum block number or block timestamp.

Type Definitions

EthAddress

Address

Index

Index in block